i dont know anymore

+++ part1_server_base/include/readgraph.h
@@ -6,13 +6,13 @@
 #include "wdigraph.h"
 
 struct Point {
-  long long lat; // latitude of the point
-  long long lon; // longitude of the point
+  ll lat; // latitude of the point
+  ll lon; // longitude of the point
 };
 
 typedef std::unordered_map<int, Point> ptMap;
 
-long long manhattan(const Point& pt1, const Point& pt2);
+ll manhattan(const Point& pt1, const Point& pt2);
 
 void readGraph(string filename, WDigraph& graph, ptMap & points);
